Sharing Styles between different versions of ReSharper.
