+++ Project/Src/AddIns/ReSharper611/CodeCleanup/Options/DocumentationOptions.cs
@@ -23,7 +23,7 @@ namespace StyleCop.ReSharper611.CodeCleanup.Options
     using System.Reflection;
     using System.Text;
 
-    using StyleCop.ReSharper611.CodeCleanup.Styles;
+    using ReSharperBase.CodeCleanup.Styles;
 
     #endregion
