+++ Project/Src/AddIns/ReSharper710/CodeCleanup/Descriptors/DocumentationDescriptor.cs
@@ -28,8 +28,9 @@ namespace StyleCop.ReSharper710.CodeCleanup.Descriptors
 
     using JetBrains.ReSharper.Feature.Services.CodeCleanup;
 
+    using ReSharperBase.CodeCleanup.Styles;
+
     using StyleCop.ReSharper710.CodeCleanup.Options;
-    using StyleCop.ReSharper710.CodeCleanup.Styles;
 
     #endregion
